add multipurpose-branch; cleanup unused methods
