+++ wmWebApp/wm.Repository/wm.Repository.csproj
@@ -57,7 +57,6 @@
     <Compile Include="GoodCategoryRepository.cs" />
     <Compile Include="GoodRepository.cs" />
     <Compile Include="GoodUnitRepository.cs" />
-    <Compile Include="MultiPurposeListGoodRepository.cs" />
     <Compile Include="MultiPurposeListRepository.cs" />
     <Compile Include="OrderGoodRepository.cs" />
     <Compile Include="OrderRepository.cs" />
