+++ wmWebApp/wm.Service/wm.Service.csproj
@@ -56,7 +56,6 @@
     <Compile Include="CalendarEvent\WarehouseKeeperCalendarEventStrategy.cs" />
     <Compile Include="EmployeeService.cs" />
     <Compile Include="CalendarEvent\CalendarEventService.cs" />
-    <Compile Include="MultiPurposeListGoodService.cs" />
     <Compile Include="GoodCategoryGoodService.cs" />
     <Compile Include="MultiPurposeListService.cs" />
     <Compile Include="GoodCategoryService.cs" />
